middleware created

+++ app/Http/Controllers/EventController.php
@@ -27,9 +27,12 @@ class EventController extends Controller
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
-    {
-        //
+    public function create(){
+
+        if (auth()->user()->role != 'Admin' && auth()->user()->role != 'Organiser') {
+            abort(403, 'Unauthorized action.');
+        }
+
         return view('events.create');
     }
 
@@ -72,9 +75,12 @@ class EventController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Event $event)
-    {
-        //
+    public function edit(Event $event){
+        
+        if (auth()->user()->role != 'Admin' && auth()->user()->role != 'Organiser') {
+            abort(403, 'Unauthorized action.');
+        }
+
         return view('events.edit', [
             'event' => $event
         ]);
@@ -106,9 +112,12 @@ class EventController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Event $event)
-    {
-        //
+    public function destroy(Event $event){
+
+        if (auth()->user()->role != 'Admin' && auth()->user()->role != 'Organiser') {
+            abort(403, 'Unauthorized action.');
+        }
+ 
         $event->delete();
         return redirect()->route('events.index')->with('message', 'Event deleted!');
     }
